+++ packages/analytics-browser/src/config.ts
@@ -102,7 +102,7 @@ export class BrowserConfig extends Config implements IBrowserConfig {
     public flushQueueSize: number = 30,
     public identityStorage: IdentityStorageType = DEFAULT_IDENTITY_STORAGE,
     public ingestionMetadata?: IngestionMetadata,
-    public instanceName?: string,
+    instanceName?: string,
     lastEventId?: number,
     lastEventTime?: number,
     public loggerProvider: ILogger = new Logger(),
@@ -140,7 +140,7 @@ export class BrowserConfig extends Config implements IBrowserConfig {
     public customEnrichment?: boolean | CustomEnrichmentOptions,
     public delayedEventsServerUrl?: string,
   ) {
-    super({ apiKey, storageProvider, transportProvider: createTransport(transport) });
+    super({ apiKey, instanceName, storageProvider, transportProvider: createTransport(transport) });
     this._cookieStorage = cookieStorage;
     this.deviceId = deviceId;
     this.lastEventId = lastEventId;